Патч в отдельный файл
